refactoring: EmojiInsertButton 컴포넌트 분리, 필터링 조건 함수 분리
