+++ frontend/guest-app/src/components/Emoji/EmojiInsertButton.js
@@ -1,0 +1,15 @@
+import IconButton from "@material-ui/core/IconButton";
+import InsertEmoticonOutlinedIcon from "@material-ui/icons/InsertEmoticonOutlined.js";
+import React from "react";
+
+function EmojiInsertButton(props) {
+	const {onClick} = props;
+
+	return (
+		<IconButton size="small" onClick={onClick}>
+			<InsertEmoticonOutlinedIcon />
+		</IconButton>
+	);
+}
+
+export default EmojiInsertButton;
